Backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.9
psycopg2-binary==2.9.5
pandas==1.5.2